requirements.txt
pdfkit
xhtml2pdf
Django==2.2.12
django-meta==1.6.1
pytz==2020.1
six==1.14.0
sqlparse==0.3.1
PyMySQL==0.9.3
django-model-utils==4.0.0
django-admin-interface==0.12.2

